+++ packages/components/src/Slider/Slider.story.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { Slider } from './Slider';
+import { Slider } from './Slider/Slider';
+import { RangeSlider } from './RangeSlider/RangeSlider';
 
 export default {
   title: 'Components/Slider',
@@ -7,27 +8,35 @@ export default {
 };
 
 export const Basic = () => {
-  const [request, setRequest] = React.useState(2);
-  const [limit, setLimit] = React.useState(3);
-  const cpuFormatter = (value) => {
-    if (value > 0 && value < 1) {
-      return value.toFixed(2);
-    }
-    if (value > 1 && value !== Infinity) {
-      return value.toFixed(1);
-    }
-    return value;
-  };
+  return (
+    <div style={{ padding: 100 }}>
+      <Slider disabled defaultValue={60} size="xs" styles={{ marginTop: 25 }} />
+      <Slider defaultValue={60} size="sm" styles={{ marginTop: 25 }} />
+      <Slider defaultValue={60} size="md" styles={{ marginTop: 25 }} />
+      <Slider defaultValue={60} size="lg" styles={{ marginTop: 25 }} />
+      <Slider defaultValue={60} size="xl" styles={{ marginTop: 25 }} />
+      <RangeSlider defaultValue={[20, 80]} styles={{ marginTop: 25 }} />
+    </div>
+  );
+};
 
-  const handleChange = (value) => {
-    setRequest(value[0]);
-    setLimit(value[1]);
-  };
-
+export const Usage = () => {
   return (
     <Slider
-      onChange={handleChange}
-      valueFormatter={cpuFormatter}
+      marks={[
+        { value: 0, label: '0' },
+        { value: 20, label: '20%' },
+        { value: 50, label: '50%' },
+        { value: 80, label: '80%' },
+        { value: 100, label: '100%' },
+      ]}
+    />
+  );
+};
+
+export const Uneven = () => {
+  return (
+    <RangeSlider
       marks={[
         { value: 0, label: '无限制', weight: 2 },
         { value: 0.2, label: 0.2, weight: 2 },
@@ -36,12 +45,41 @@ export const Basic = () => {
         { value: 2, label: 2, weight: 2 },
         { value: 3, label: 3, weight: 1 },
         { value: 4, label: 4 },
-        { value: Infinity, label: '无限制' },
       ]}
-      value={[request || 0, limit || Infinity]}
-      defaultValue={10}
-      unit="MiB"
-      hasTooltip
+      defaultValue={[2, 3]}
+      minRange={0.01}
+      decimals={2}
     />
+  );
+};
+
+export const Controlled = () => {
+  // @ts-ignore
+  const marks = [
+    { value: 0, label: '0' },
+    { value: 20, label: '20%' },
+    { value: 50, label: '50%' },
+    { value: 80, label: '80%' },
+    { value: 100, label: '100%' },
+  ];
+  return <RangeSlider defaultValue={[20, 80]} marks={marks} />;
+};
+
+export const ControlLabel = () => {
+  return (
+    <div style={{ padding: 100 }}>
+      <Slider label={null} styles={{ marginTop: 25 }} />
+
+      <Slider label={(value) => `${value} °C`} styles={{ marginTop: 25 }} />
+
+      <Slider labelAlwaysOn styles={{ marginTop: 25 }} />
+
+      <Slider
+        styles={{ marginTop: 25 }}
+        labelTransition="skew-down"
+        labelTransitionDuration={150}
+        labelTransitionTimingFunction="ease"
+      />
+    </div>
   );
 };